feat(lvm-systemd): introducing the lvm-systemd module
Reduce the number of lvm commands, and the amount of
lvm device scanning, to improve the efficiency and
speed of root LV activation.  Previously, LVs were
activated upon the arrival of every PV, and these
activations would fail until all PVs had arrived.
The new module keeps track of PVs that have arrived,
and only activates LVs when they are complete, so
activation is only run when it will succeed.

The previous 90lvm module is renamed 90lvm-initqueue.

The new systemd-based module is named 90lvm-systemd.

The new 90lvm meta module chooses between them, based
on features available in pvscan.

In the new 90lvm-systemd module:

- A new udev rule is run for each lvm device (PV) that
  appears, and the udev rule runs pvscan directly on each
  device.  pvscan reads only the specified device, which
  avoids excessive device scanning when many PVs appear.

- New pvscan options inform the udev rule when an LV is
  complete and ready to be activated, based on the arrival
  of the given device.

- pvscan is run via a shell wrapper that compares complete
  LV names to those named by rd.lvm.lv, so that only LVs
  named by rd.lvm.lv are activated (this avoids unnecessary
  activations that may delay startup.)

- Based on the points above, the udev rule knows when a
  required LV can be successfully activated, and it will
  run a single lvm command to activate each required LV.
  This avoids running many unnecessary (premature) activation
  commands that can delay startup.

- systemd-run is used to run the activation commands because
  the lvm command may take longer to run than udev permits.

- In case rd.lvm.vg is used (in place of, or in addition to
  rd.lvm.lv), adjust the points agove to refer to VG.

- This method is very similar to the new autoactivation
  method that lvm is adopting when run from root.